refactor(providers): change provider file location
