--- src/modules/user/providers/HashProvider/implementantions/BCryptHashProvider.ts
@@ -1,12 +1,0 @@
-import { compare, hash } from 'bcryptjs';
-import IHashProvider from '../models/IHashProvider';
-
-export default class BCryptHashProvider implements IHashProvider {
-  public async generateHash(payload: string): Promise<string> {
-    return hash(payload, 8);
-  }
-
-  public async compareHash(payload: string, hashed: string): Promise<boolean> {
-    return compare(payload, hashed);
-  }
-}
