--- src/modules/user/providers/HashProvider/models/IHashProvider.ts
@@ -1,4 +1,0 @@
-export default interface IHashProvider {
-  generateHash(payload: string): Promise<string>;
-  compareHash(payload: string, hashed: string): Promise<boolean>;
-}
