--- src/modules/user/providers/index.ts
@@ -1,6 +1,0 @@
-import { container } from 'tsyringe';
-
-import IHashProvider from './HashProvider/models/IHashProvider';
-import BCryptHashProvider from './HashProvider/implementantions/BCryptHashProvider';
-
-container.registerSingleton<IHashProvider>('HashProvider', BCryptHashProvider);
