+++ src/modules/user/services/CreateUserService.ts
@@ -5,7 +5,7 @@ import AppError from '@shared/infra/http/errors/AppError';
 import { azureCreate } from '@shared/infra/azure/imageStorage/imageUpload';
 import User from '../infra/typeorm/entities/User';
 import IUserRepository from '../repositories/IUserRepository';
-import IHashProvider from '../providers/HashProvider/models/IHashProvider';
+import IHashProvider from '../../../shared/providers/HashProvider/models/IHashProvider';
 
 export interface IRequest {
   name: string;
