+++ src/shared/container/index.ts
@@ -1,6 +1,6 @@
 import { container } from 'tsyringe';
 
-import '@modules/user/providers';
+import '@shared/providers';
 
 import UsersRepository from '@modules/user/infra/typeorm/repositories/UserRepository';
 import IUsersRepository from '@modules/user/repositories/IUserRepository';
