+++ src/shared/providers/HashProvider/models/IHashProvider.ts
